feat: Enhance Navbar
Changed the pages array to an array of objects with two keys: label and link.
Wrapped the navigation links inside the Link component from react-router-dom, using the link key from each object in the pages array as the 'to' prop of Link.
Removed the MenuItem component from the Material-UI menu in favor of using a Link component inside a Button component.
Changed the color of the AppBar background and added padding to it.
Removed the borderBottom property from the AppBar.
Removed the boxShadow property from the AppBar.